Working http3 selection meanings into the code, centralizuing ALPN selections for vtls backends.
    - ALPN selection for SSL/SSL-PROXY filters in one vtls set of functions, replaces
      copy of logic in all tls backends.
    - standardized the infof logging of offered ALPNs
    - ALPN negotiated: have common function for all backends that sets alpn proprty
      and connection related things based on the negotiated protocol (or lack thereof).

    tests-https:

    - Adding test cases for h3 eyeballing: --http3 and --http3-only
        - --http3 skipped for now as implementation missing
        - adding stop/start checks for nghttpx
        - adding nghttpx start fixtures to test that need it, as
          we have now tests that stop it
    - Make HTTPS and QUIC port the same in tests/tests-httpd

+++ lib/cf-socket.c
@@ -1130,8 +1130,8 @@ static ssize_t cf_socket_send(struct Curl_cfilter *cf, struct Curl_easy *data,
 
   DEBUGASSERT(data->conn == cf->conn);
   nwritten = Curl_send_plain(data, cf->sockindex, buf, len, err);
-  DEBUGF(LOG_CF(data, cf, "send(len=%zu) -> %d, err=%d",
-                len, (int)nwritten, *err));
+  DEBUGF(LOG_CF(data, cf, "send(len=%zu) -> %zd, err=%d",
+                len, nwritten, *err));
   return nwritten;
 }
 
@@ -1142,8 +1142,8 @@ static ssize_t cf_socket_recv(struct Curl_cfilter *cf, struct Curl_easy *data,
 
   DEBUGASSERT(data->conn == cf->conn);
   nread = Curl_recv_plain(data, cf->sockindex, buf, len, err);
-  DEBUGF(LOG_CF(data, cf, "recv(len=%zu) -> %d, err=%d", len, (int)nread,
-                *err));
+  DEBUGF(LOG_CF(data, cf, "recv(len=%zu) -> %zd, err=%d",
+                len, nread, *err));
   return nread;
 }
